+++ Package.swift
@@ -18,6 +18,10 @@ let package = Package(
             dependencies: [
                 .product(name: "ArgumentParser", package: "swift-argument-parser")
             ]
+        ),
+        .testTarget(
+            name: "webwrapTests",
+            dependencies: ["webwrap"]
         )
     ]
 )